Working on configuration parsing.
Thinking something along the lines of a std::accumulate with each transform
returning a new options struct and the final result is the set of valid options.
There's probably something missing here as there will probably need to be steps in
between that transforms to optionals, then filters to remove empties.